modified profile page and fetched data from db

+++ backend/project/app1/models.py
@@ -11,8 +11,8 @@ class EmergencyContact(models.Model):
 class SOSRequest(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     emergency_type = models.CharField(max_length=20)
-    latitude = models.DecimalField(max_digits=9, decimal_places=6)
-    longitude = models.DecimalField(max_digits=9, decimal_places=6)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     message = models.TextField(blank=True, null=True)
     created_at = models.DateTimeField(auto_now_add=True)
